updated data types

+++ Modules/Sales/Traits/HasShippingCalculation.php
@@ -30,11 +30,11 @@ trait HasShippingCalculation
 				else {
 					$order->order_taxes->map( function ($order_tax) use (&$total_shipping_amount) {
 						$order_item_total_amount = $order_tax->order_tax_items->filter(fn ($order_tax_item) => ($order_tax_item->tax_item_type == "product"))->map( function ($order_item) use ($order_tax, &$total_shipping_amount) {
-							$amount = (($order_tax->percent/100) * $order_item->amount);
+							$amount = (float) (($order_tax->percent/100) * $order_item->amount);
 							$total_shipping_amount += $amount;
 							$data = [
 								"tax_id" => $order_tax->id,
-								"tax_percent" => $order_tax->percent,
+								"tax_percent" => (float) $order_tax->percent,
 								"amount" => $amount,
 								"tax_item_type" => "shipping"
 							];
